[style] login class 분리
